chat_app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
jinja2
google-generativeai
google-cloud-texttospeech
transformers==4.26.1
